// src/app/model/CabeceraSac.ts
export class CabeceraSac {
    idcabeSac: number;
    numSac: number;
    codCcosto: number;
    idTdr: number;
    anio: string;
    fechaSac: any;
    descripcion: string;
    coordinador: string;
    telefono: string;
   cabeceraDetalleSacsByIdcabeSac: number;
    centroCostoByCodCcosto: number;
    tdrByIdTdr: number;
}